chemical_project/backend/requirement.txt
django
djangorestframework
django-cors-headers
pandas
